+++ server/src/loaders/index.js
@@ -1,11 +1,11 @@
 const dependencyInjectionLoader = require('./dependency-injector');
 const expressLoader = require('./express');
 const socketLoader = require('./socket');
-const sessionExpiredLoader = require('./session-expired');
+const jobsLoader = require('./jobs');
 
 module.exports = (app, io) => {
   dependencyInjectionLoader();
   expressLoader(app);
   socketLoader(io);
-  sessionExpiredLoader();
+  jobsLoader();
 };
