--- server/src/loaders/session-expired.js
@@ -1,11 +1,0 @@
-const { Container } = require('typedi');
-
-module.exports = () => {
-  const sessionManager = Container.get('sessionManager');
-  const pushService = Container.get('pushService');
-
-  sessionManager.onSessionExpired(sessionId => {
-    sessionManager.terminateSession(sessionId);
-    pushService.removeSubscription(sessionId);
-  });
-};
