+++ server/src/services/session-manager.js
@@ -1,7 +1,9 @@
-const { session } = require('../config');
 const { getCurrentDate } = require('../utils');
-const CHECK_SESSION_EXPIRED_INTERVAL = 60 * 1000;
 const sessions = new Map();
+
+function getSessions() {
+  return sessions;
+}
 
 function hasSession(sessionId) {
   return sessions.has(sessionId);
@@ -48,21 +50,8 @@ function markSessionAsInactive(sessionId) {
   });
 }
 
-function onSessionExpired(callback) {
-  setInterval(() => {
-    Array.from(sessions)
-      .filter(([_, session]) => !session.isActive && session.disconnectedAt)
-      .filter(([_, session]) => hasSessionExpired(session.disconnectedAt))
-      .forEach(([id, _]) => callback(id));
-  }, CHECK_SESSION_EXPIRED_INTERVAL);
-}
-
-function hasSessionExpired(disconnectedAt) {
-  const timeSinceDisconnect = new Date() - new Date(disconnectedAt);
-  return timeSinceDisconnect > session.maxDisconnectTime;
-}
-
 module.exports = {
+  getSessions,
   hasSession,
   startSession,
   terminateSession,
@@ -70,6 +59,5 @@ module.exports = {
   getActiveUsers,
   registerUser,
   isUsernameAvailable,
-  markSessionAsInactive,
-  onSessionExpired
+  markSessionAsInactive
 };
